Move Roles entity to user models
It makes more sense in the distribution of the app
