+++ app/user/models.py
@@ -3,7 +3,7 @@ from __future__ import unicode_literals
 from app import db
 from app.util import now, enum, uuid
 from app.sql import ChoiceType
-from app.constants import Genders
+from app.constants import Genders, Roles
 
 from flask.ext.login import UserMixin
 from passlib.hash import sha256_crypt
@@ -57,6 +57,26 @@ class User(db.Model, UserMixin):
 
     def get_id(self):
         return unicode(self.username)
+
+    def is_active(self):
+        # Only administrators are allowed to login
+        return Grant.check_grant(self, Roles.ADMIN)
+
+
+class Grant(db.Model):
+    __tablename__ = 'grants'
+
+    id = db.Column(db.Integer, primary_key=True)
+    created = db.Column(db.DateTime, default=now)
+    modified = db.Column(db.DateTime, default=now, onupdate=now)
+    role = db.Column(ChoiceType(Roles), nullable=False)
+    user_id = db.Column(db.Integer, db.ForeignKey('users.id'), nullable=False)
+
+    user = db.relationship('User')
+
+    @classmethod
+    def check_grant(cls, user, role):
+        return cls.query.filter(Grant.user == user, Grant.role == role).first() is not None
 
 
 class UserDetails(db.Model):
